This is the Inventory management Information System software- Groupwork done by 3 students
The main features of this software 
is to show the available list of Appliances so that staff should not roam around the 
store searching if the Appliances is available or not. So,By Using this software 
Inventory Information System can organize data more efficiently as they can store 
information of such as ID, Name of product, price, price range i.e. Low,Medum or High
Company name etc. 
As well, we can also access the recorded information easily as there are two search 
buttons named one is search by price which search product with price and give result 
of product name and ID of that price